+++ AssimpNet.Test/TestHelper.cs
@@ -20,10 +20,10 @@
 * THE SOFTWARE.
 */
 
+using NUnit.Framework;
 using System;
 using System.IO;
 using System.Reflection;
-using NUnit.Framework;
 using TK = OpenTK;
 
 namespace Assimp.Test
@@ -33,8 +33,8 @@ namespace Assimp.Test
     /// </summary>
     public static class TestHelper
     {
-        public const float DEFAULT_TOLERANCE = 0.000001f;
-        public static float Tolerance = DEFAULT_TOLERANCE;
+        public const double DEFAULT_TOLERANCE = 0.000001f;
+        public static double Tolerance = DEFAULT_TOLERANCE;
 
         private static string m_rootPath = null;
 
@@ -42,7 +42,7 @@ namespace Assimp.Test
         {
             get
             {
-                if(m_rootPath == null)
+                if (m_rootPath == null)
                 {
                     /*
                     Assembly entryAssembly = Assembly.GetEntryAssembly();
@@ -83,33 +83,33 @@ namespace Assimp.Test
             Assert.That(actual, Is.EqualTo(expected).Within(Tolerance), msg);
         }
 
-        public static void AssertEquals(float x, float y, Vector2D v)
+        public static void AssertEquals(double x, double y, Vector2D v)
         {
             AssertEquals(x, v.X);
             AssertEquals(y, v.Y);
         }
 
-        public static void AssertEquals(float x, float y, Vector2D v, String msg)
+        public static void AssertEquals(double x, double y, Vector2D v, String msg)
         {
             AssertEquals(x, v.X, msg + $" => checking X component ({x} == {v.X}");
             AssertEquals(y, v.Y, msg + $" => checking Y component ({y} == {v.Y}");
         }
 
-        public static void AssertEquals(float x, float y, float z, Vector3D v)
+        public static void AssertEquals(double x, double y, double z, Vector3D v)
         {
             AssertEquals(x, v.X);
             AssertEquals(y, v.Y);
             AssertEquals(z, v.Z);
         }
 
-        public static void AssertEquals(float x, float y, float z, Vector3D v, String msg)
+        public static void AssertEquals(double x, double y, double z, Vector3D v, String msg)
         {
             AssertEquals(x, v.X, msg + $" => checking X component ({x} == {v.X}");
             AssertEquals(y, v.Y, msg + $" => checking Y component ({y} == {v.Y}");
             AssertEquals(z, v.Z, msg + $" => checking Z component ({z} == {v.Z}");
         }
 
-        public static void AssertEquals(float r, float g, float b, float a, Color4D c)
+        public static void AssertEquals(double r, double g, double b, double a, Color4D c)
         {
             AssertEquals(r, c.R);
             AssertEquals(g, c.G);
@@ -117,7 +117,7 @@ namespace Assimp.Test
             AssertEquals(a, c.A);
         }
 
-        public static void AssertEquals(float r, float g, float b, float a, Color4D c, String msg)
+        public static void AssertEquals(double r, double g, double b, double a, Color4D c, String msg)
         {
             AssertEquals(r, c.R, msg + String.Format(" => checking R component ({0} == {1}", r, c.R));
             AssertEquals(g, c.G, msg + String.Format(" => checking G component ({0} == {1}", g, c.G));
@@ -125,21 +125,21 @@ namespace Assimp.Test
             AssertEquals(a, c.A, msg + String.Format(" => checking A component ({0} == {1}", a, c.A));
         }
 
-        public static void AssertEquals(float r, float g, float b, Color3D c)
+        public static void AssertEquals(double r, double g, double b, Color3D c)
         {
             AssertEquals(r, c.R);
             AssertEquals(g, c.G);
             AssertEquals(b, c.B);
         }
 
-        public static void AssertEquals(float r, float g, float b, Color3D c, String msg)
+        public static void AssertEquals(double r, double g, double b, Color3D c, String msg)
         {
             AssertEquals(r, c.R, msg + String.Format(" => checking R component ({0} == {1}", r, c.R));
             AssertEquals(g, c.G, msg + String.Format(" => checking G component ({0} == {1}", g, c.G));
             AssertEquals(b, c.B, msg + String.Format(" => checking B component ({0} == {1}", b, c.B));
         }
 
-        public static void AssertEquals(float x, float y, float z, float w, Quaternion q, string msg)
+        public static void AssertEquals(double x, double y, double z, double w, Quaternion q, string msg)
         {
             AssertEquals(x, q.X, msg + $" => checking X component ({x} == {q.X}");
             AssertEquals(y, q.Y, msg + $" => checking Y component ({y} == {q.Y}");
@@ -178,10 +178,10 @@ namespace Assimp.Test
         public static void AssertEquals(TK.Matrix4 tkM, Matrix4x4 mat, string msg)
         {
             //Note: OpenTK 4x4 matrix is a row-vector matrix, so compare rows to AssimpNet Matrix4x4 columns
-            AssertEquals(tkM.Row0, new TK.Vector4(mat.A1, mat.B1, mat.C1, mat.D1), msg + " => checking first column vector");
-            AssertEquals(tkM.Row1, new TK.Vector4(mat.A2, mat.B2, mat.C2, mat.D2), msg + " => checking second column vector");
-            AssertEquals(tkM.Row2, new TK.Vector4(mat.A3, mat.B3, mat.C3, mat.D3), msg + " => checking third column vector");
-            AssertEquals(tkM.Row3, new TK.Vector4(mat.A4, mat.B4, mat.C4, mat.D4), msg + " => checking third column vector");
+            AssertEquals(tkM.Row0, new TK.Vector4((float)mat.A1, (float)mat.B1, (float)mat.C1, (float)mat.D1), msg + " => checking first column vector");
+            AssertEquals(tkM.Row1, new TK.Vector4((float)mat.A2, (float)mat.B2, (float)mat.C2, (float)mat.D2), msg + " => checking second column vector");
+            AssertEquals(tkM.Row2, new TK.Vector4((float)mat.A3, (float)mat.B3, (float)mat.C3, (float)mat.D3), msg + " => checking third column vector");
+            AssertEquals(tkM.Row3, new TK.Vector4((float)mat.A4, (float)mat.B4, (float)mat.C4, (float)mat.D4), msg + " => checking third column vector");
         }
     }
 }
